backend/requirements.txt
# Django Core
Django==5.0
django-ninja==1.1.0

# Database
psycopg2-binary==2.9.9

# Configuration
python-decouple==3.8

# CORS
django-cors-headers==4.3.1

# Task Queue
celery==5.3.4
redis==5.0.1
django-celery-beat>=2.6.0

# Authentication
PyJWT==2.8.0

# Security
cryptography==41.0.7

# API & HTTP
requests==2.31.0
urllib3==2.1.0

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0

# Code Quality
black==23.12.1
isort==5.13.2
flake8==6.1.0

# Production Server
gunicorn==21.2.0
